gamium/python/pytest/requirements.txt
Appium-Python-Client==3.1.0
selenium==4.15.2
python-dotenv==1.0.0
dogu-device-client==1.0.0
pytest==7.4.0
pytest-dogu-sdk==1.0.4
gamium==2.0.7
